requirements.txt
# Lexora AI
# Enterprise Knowledge Intelligence Platform

## Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

## Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

## Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

## LangChain & AI
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
openai==1.10.0
tiktoken==0.5.2
huggingface-hub==0.20.3

## Document Processing
pypdf==3.17.4
pymupdf==1.23.8
python-docx==1.1.0

## Caching & Async
redis==5.0.1
celery==5.3.6
httpx==0.26.0

## Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0

## Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3
aiofiles==23.2.1
